Move LayoutThreadData to query.rs.
